Components: Move HighlightCards to client
