+++ .storybook/main.js
@@ -7,6 +7,7 @@ const storybookConfig = storybookDefaultConfig( {
 		'../client/components/**/*.stories.{js,jsx,tsx}',
 		'../client/blocks/**/*.stories.{js,jsx,tsx}',
 		'../client/dashboard/**/*.stories.{js,jsx,tsx}',
+		'../client/my-sites/stats/components/**/*.stories.{js,jsx,tsx}',
 
 		'../packages/design-picker/src/**/*.stories.{ts,tsx}',
 		'../packages/domains-table/src/**/*.stories.{js,jsx,ts,tsx}',
