+++ client/my-sites/earn/customers/customer/customer-stats.tsx
@@ -3,7 +3,7 @@ import { formatCurrency } from '@automattic/number-formatters';
 import { payment, chartBar } from '@wordpress/icons';
 import { Subscriber } from '../../types';
 
-import '@automattic/components/src/highlight-cards/style.scss';
+import 'calypso/my-sites/stats/components/highlight-cards/style.scss';
 
 type CustomerStatsProps = {
 	customer: Subscriber;
