+++ client/my-sites/stats/components/highlight-cards/variables.scss
