+++ client/my-sites/stats/post-detail-highlights-section/style.scss
@@ -1,6 +1,6 @@
 @use "sass:math";
 @import "@wordpress/base-styles/breakpoints";
-@import "@automattic/components/src/highlight-cards/variables";
+@import "calypso/my-sites/stats/components/highlight-cards/variables";
 @import "@wordpress/base-styles/mixins";
 
 .stats__post-detail-highlights-section {
