+++ client/my-sites/stats/sections/all-time-highlights-section/index.tsx
@@ -1,8 +1,4 @@
 import { Card, ComponentSwapper, DotPager } from '@automattic/components';
-import {
-	formatPercentage,
-	percentCalculator,
-} from '@automattic/components/src/highlight-cards/lib/numbers';
 import { eye } from '@automattic/components/src/icons';
 import { formatNumber, formatNumberCompact } from '@automattic/number-formatters';
 import { Icon, people, postContent, starEmpty, commentContent } from '@wordpress/icons';
@@ -11,6 +7,10 @@ import { useTranslate } from 'i18n-calypso';
 import React, { useMemo } from 'react';
 import QueryPosts from 'calypso/components/data/query-posts';
 import QuerySiteStats from 'calypso/components/data/query-site-stats';
+import {
+	formatPercentage,
+	percentCalculator,
+} from 'calypso/my-sites/stats/components/highlight-cards/lib/numbers';
 import { useSelector } from 'calypso/state';
 import { getCurrentUserLocale } from 'calypso/state/current-user/selectors';
 import {
