+++ client/my-sites/stats/sections/all-time-highlights-section/style.scss
@@ -1,6 +1,6 @@
 @import "@wordpress/base-styles/breakpoints";
 @import "@automattic/typography/styles/variables";
-@import "@automattic/components/src/highlight-cards/variables";
+@import "calypso/my-sites/stats/components/highlight-cards/variables";
 
 .stats__all-time-highlights-section {
 	.highlight-cards {
