+++ client/my-sites/stats/sections/annual-highlights-section/style.scss
@@ -1,5 +1,5 @@
 @import "@automattic/typography/styles/variables";
-@import "@automattic/components/src/highlight-cards/variables";
+@import "calypso/my-sites/stats/components/highlight-cards/variables";
 
 .highlight-year-navigation {
 	.stats-navigation-arrows {
