+++ client/my-sites/stats/sections/posting-activity-section/style.scss
@@ -1,7 +1,7 @@
 @import "@automattic/typography/styles/variables";
 @import "@wordpress/base-styles/breakpoints";
 @import "@automattic/components/src/styles/mixins";
-@import "@automattic/components/src/highlight-cards/variables";
+@import "calypso/my-sites/stats/components/highlight-cards/variables";
 
 $font-label: "SF Pro Text", $sans;
 $mobile-layout-breakpoint: $break-small;
