+++ client/my-sites/stats/stats-subscribers-chart-section/style.scss
@@ -1,5 +1,5 @@
 @import "@wordpress/base-styles/breakpoints";
-@import "@automattic/components/src/highlight-cards/variables";
+@import "calypso/my-sites/stats/components/highlight-cards/variables";
 
 .subscribers-section {
 	margin: 1em 0;
