+++ client/my-sites/stats/wordads/earnings.scss
@@ -1,6 +1,6 @@
 @import "@automattic/components/src/styles/mixins";
 @import "@automattic/typography/styles/variables";
-@import "@automattic/components/src/highlight-cards/variables";
+@import "calypso/my-sites/stats/components/highlight-cards/variables";
 
 
 .ads__table-header {
